fix: adress changes from PR comments, split files for modules, adjustments on formik components
