+++ src/pages/sponsors-global/page-templates/page-template-modules-form.js
@@ -7,193 +7,18 @@ import {
   AccordionSummary,
   AccordionDetails,
   Box,
-  Grid2,
   IconButton,
-  Typography,
-  MenuItem,
-  InputLabel,
-  Divider
+  Typography
 } from "@mui/material";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import DeleteIcon from "@mui/icons-material/Delete";
 import UnfoldMoreIcon from "@mui/icons-material/UnfoldMore";
-import MuiFormikTextField from "../../../components/mui/formik-inputs/mui-formik-textfield";
-import MuiFormikSelect from "../../../components/mui/formik-inputs/mui-formik-select";
-import MuiFormikDatepicker from "../../../components/mui/formik-inputs/mui-formik-datepicker";
-import MuiFormikRadioGroup from "../../../components/mui/formik-inputs/mui-formik-radio-group";
 import DragAndDropList from "../../../components/mui/dnd-list";
 import showConfirmDialog from "../../../components/mui/showConfirmDialog";
-import {
-  PAGES_MODULE_KINDS,
-  PAGE_MODULES_MEDIA_TYPES
-} from "../../../utils/constants";
-import FormikTextEditor from "../../../components/inputs/formik-text-editor";
-import MuiFormikUpload from "../../../components/mui/formik-inputs/mui-formik-upload";
-
-const InfoModule = ({ baseName, index }) => {
-  const buildFieldName = (field) => `${baseName}[${index}].${field}`;
-
-  return (
-    <Grid2 container spacing={2} size={12}>
-      <InputLabel htmlFor={buildFieldName("content")}>
-        {T.translate("page_template_list.page_crud.info_content")}
-      </InputLabel>
-      <Grid2 size={12}>
-        <Box width="100%">
-          <FormikTextEditor
-            name={buildFieldName("content")}
-            fullWidth
-            margin="none"
-          />
-        </Box>
-      </Grid2>
-    </Grid2>
-  );
-};
-
-const DocumentDownloadModule = ({ baseName, index }) => {
-  const buildFieldName = (field) => `${baseName}[${index}].${field}`;
-
-  return (
-    <Grid2 container spacing={2} size={12}>
-      <InputLabel htmlFor={buildFieldName("name")}>
-        {T.translate("page_template_list.page_crud.document_name")}
-      </InputLabel>
-      <Grid2 size={12}>
-        <MuiFormikTextField
-          name={buildFieldName("name")}
-          fullWidth
-          margin="none"
-        />
-      </Grid2>
-      <InputLabel htmlFor={buildFieldName("description")}>
-        {T.translate("page_template_list.page_crud.description")}
-      </InputLabel>
-      <Grid2 size={12}>
-        <MuiFormikTextField
-          name={buildFieldName("description")}
-          fullWidth
-          multiline
-          rows={2}
-          margin="none"
-        />
-      </Grid2>
-      <InputLabel htmlFor={buildFieldName("external_url")}>
-        {T.translate("page_template_list.page_crud.external_url")}
-      </InputLabel>
-      <Grid2 size={12}>
-        <MuiFormikTextField
-          name={buildFieldName("external_url")}
-          placeholder="Link"
-          fullWidth
-          margin="none"
-        />
-      </Grid2>
-      <Grid2 size={12}>
-        <MuiFormikUpload
-          id="document-module-upload"
-          name={buildFieldName("file")}
-          singleFile
-        />
-      </Grid2>
-    </Grid2>
-  );
-};
-
-const MediaRequestModule = ({ baseName, index }) => {
-  const { values } = useFormikContext();
-  const buildFieldName = (field) => `${baseName}[${index}].${field}`;
-
-  const mediaType =
-    getIn(values, buildFieldName("type")) || PAGE_MODULES_MEDIA_TYPES.FILE;
-
-  const mediaTypeOptions = [
-    {
-      value: PAGE_MODULES_MEDIA_TYPES.FILE,
-      label: T.translate("page_template_list.page_crud.upload_file")
-    },
-    {
-      value: PAGE_MODULES_MEDIA_TYPES.TEXT,
-      label: T.translate("page_template_list.page_crud.text_input")
-    }
-  ];
-
-  return (
-    <Grid2 container spacing={2} size={12}>
-      <Grid2 size={12}>
-        <MuiFormikRadioGroup
-          name={buildFieldName("type")}
-          options={mediaTypeOptions}
-          row
-          margin="none"
-        />
-      </Grid2>
-      <Grid2 size={12}>
-        <Divider sx={{ mx: -2 }} />
-      </Grid2>
-      <Grid2 size={6}>
-        <InputLabel htmlFor={buildFieldName("name")}>
-          {T.translate("page_template_list.page_crud.name")}
-        </InputLabel>
-        <MuiFormikTextField
-          name={buildFieldName("name")}
-          fullWidth
-          margin="none"
-        />
-      </Grid2>
-      <Grid2 size={6}>
-        <InputLabel htmlFor={buildFieldName("upload_deadline")}>
-          {T.translate("page_template_list.page_crud.upload_deadline")}
-        </InputLabel>
-        <MuiFormikDatepicker
-          name={buildFieldName("upload_deadline")}
-          margin="none"
-        />
-      </Grid2>
-
-      {mediaType === PAGE_MODULES_MEDIA_TYPES.FILE && (
-        <>
-          <Grid2 size={6}>
-            <InputLabel htmlFor={buildFieldName("max_file_size")}>
-              {T.translate("page_template_list.page_crud.max_file_size")}
-            </InputLabel>
-            <MuiFormikTextField
-              name={buildFieldName("max_file_size")}
-              type="number"
-              margin="none"
-              fullWidth
-            />
-          </Grid2>
-          <Grid2 size={6}>
-            <InputLabel htmlFor={buildFieldName("max_file_size")}>
-              {T.translate("page_template_list.page_crud.allowed_formats")}
-            </InputLabel>
-            <MuiFormikSelect
-              name={buildFieldName("file_type_id")}
-              margin="none"
-              fullWidth
-            >
-              <MenuItem value={1}>PDF</MenuItem>
-            </MuiFormikSelect>
-          </Grid2>
-        </>
-      )}
-
-      <Grid2 size={12}>
-        <InputLabel htmlFor={buildFieldName("max_file_size")}>
-          {T.translate("page_template_list.page_crud.description")}
-        </InputLabel>
-        <MuiFormikTextField
-          name={buildFieldName("description")}
-          fullWidth
-          multiline
-          margin="none"
-          rows={2}
-        />
-      </Grid2>
-    </Grid2>
-  );
-};
+import { PAGES_MODULE_KINDS } from "../../../utils/constants";
+import InfoModule from "./modules/page-template-info-module";
+import DocumentDownloadModule from "./modules/page-template-document-download-module";
+import MediaRequestModule from "./modules/page-template-media-request-module";
 
 const PageModules = ({ name = "modules" }) => {
   const { values, setFieldValue } = useFormikContext();
