+++ src/pages/sponsors/sponsor-pages-list-page/index.js
@@ -24,6 +24,7 @@ import {
 } from "@mui/material";
 import AddIcon from "@mui/icons-material/Add";
 import { getSponsorPages } from "../../../actions/sponsor-pages-actions";
+import { getSponsorForm } from "../../../actions/sponsor-forms-actions";
 import CustomAlert from "../../../components/mui/custom-alert";
 import MuiTable from "../../../components/mui/table/mui-table";
 import GlobalPagePopup from "./components/global-page/global-page-popup";
@@ -144,6 +145,7 @@ const SponsorPagesListPage = ({
               control={
                 <Checkbox
                   onChange={handleHideArchivedForms}
+                  checked={hideArchived}
                   inputProps={{
                     "aria-label": T.translate("sponsor_pages.hide_archived")
                   }}
@@ -219,5 +221,6 @@ const mapStateToProps = ({ sponsorPagesListState }) => ({
 });
 
 export default connect(mapStateToProps, {
-  getSponsorPages
+  getSponsorPages,
+  getSponsorForm
 })(SponsorPagesListPage);
